Moved the legacy code to legacy.py to help better organize things
