+++ addon.py
@@ -19,47 +19,12 @@ import xbmcplugin
 
 from resources.lib import util
 from resources.lib import player_config
-from resources.lib import events
 from resources.lib import adobe_activate_api
 from resources.lib.globals import selfAddon, defaultlive, defaultreplay, defaultupcoming, defaultimage, defaultfanart, translation, pluginhandle, LOG_LEVEL
+from resources.lib.constants import *
+from resources.lib.addon_util import *
 
-OLD_LISTING_MODE = 'OLD_LISTING_MODE'
-LIVE_EVENTS_MODE = 'LIVE_EVENTS'
-PLAY_MODE = 'PLAY'
-PLAY_ITEM_MODE = 'PLAY_ITEM'
-PLAY_TV_MODE = 'PLAY_TV'
-LIST_SPORTS_MODE = 'LIST_SPORTS'
-INDEX_SPORTS_MODE = 'INDEX_SPORTS'
-UPCOMING_MODE = 'UPCOMING'
-AUTHENTICATE_MODE = 'AUTHENTICATE'
-AUTHENTICATION_DETAILS_MODE = 'AUTHENTICATION_DETAILS'
-CATEGORY_SHELF_MODE = 'CATEGORY_SHELF'
-CATEGORY_SHOWCASE_MODE = 'CATEGORY_SHOWCASE'
-CATEGORY_SPORTS_MODE = 'CATEGORY_SPORTS'
-CATEGORY_CHANNELS_MODE = 'CATEGORY_CHANNELS'
-NETWORK_ID = 'NETWORK_ID'
-EVENT_ID = 'EVENT_ID'
-SIMULCAST_AIRING_ID = 'SIMULCAST_AIRING_ID'
-SESSION_URL = 'SESSION_URL'
-DESKTOP_STREAM_SOURCE = 'DESKTOP_STREAM_SOURCE'
-NETWORK_NAME = 'NETWORK_NAME'
-EVENT_NAME = 'EVENT_NAME'
-EVENT_GUID = 'EVENT_GUID'
-EVENT_PARENTAL_RATING = 'EVENT_PARENTAL_RATING'
-SHELF_ID = 'SHELF_ID'
-SHOWCASE_URL = 'SHOWCASE_URL'
-SHOWCASE_NAV_ID = 'SHOWCASE_NAV_ID'
-PLAYBACK_URL = 'PLAYBACK_URL'
-
-ESPN_URL = 'ESPN_URL'
-MODE = 'MODE'
-SPORT = 'SPORT'
-
-BAM_NS = '{http://services.bamnetworks.com/media/types/2.1}'
-
-# Taken from https://espn.go.com/watchespn/player/config
-ESPN3_ID = 'n360'
-SECPLUS_ID = 'n323'
+from resources.lib import legacy
 
 TAG = 'ESPN3: '
 
@@ -334,217 +299,6 @@ def INDEX_TV_SHELF(stash_json, item, upcoming):
         authurl[EVENT_PARENTAL_RATING] = mpaa
     addLink(ename.encode('iso-8859-1'), authurl, fanart, fanart, infoLabels=infoLabels)
 
-def CATEGORIES():
-    include_premium = adobe_activate_api.is_authenticated()
-    channel_list = events.get_channel_list(include_premium)
-    curdate = datetime.utcnow()
-    upcoming = int(selfAddon.getSetting('upcoming'))+1
-    days = (curdate+timedelta(days=upcoming)).strftime("%Y%m%d")
-    addDir(translation(30029),
-           dict(ESPN_URL=events.get_live_events_url(channel_list), MODE=LIVE_EVENTS_MODE),
-           defaultlive)
-    addDir(translation(30030),
-           dict(ESPN_URL=events.get_upcoming_events_url(channel_list) + '&endDate=' + days + '&startDate=' + curdate.strftime("%Y%m%d"), MODE=LIST_SPORTS_MODE),
-           defaultupcoming)
-    enddate = '&endDate='+ (curdate+timedelta(days=1)).strftime("%Y%m%d")
-    replays1 = [5,10,15,20,25]
-    replays1 = replays1[int(selfAddon.getSetting('replays1'))]
-    start1 = (curdate-timedelta(days=replays1)).strftime("%Y%m%d")
-    replays2 = [10,20,30,40,50]
-    replays2 = replays2[int(selfAddon.getSetting('replays2'))]
-    start2 = (curdate-timedelta(days=replays2)).strftime("%Y%m%d")
-    replays3 = [30,60,90,120]
-    replays3 = replays3[int(selfAddon.getSetting('replays3'))]
-    start3 = (curdate-timedelta(days=replays3)).strftime("%Y%m%d")
-    replays4 = [60,90,120,240]
-    replays4 = replays4[int(selfAddon.getSetting('replays4'))]
-    start4 = (curdate-timedelta(days=replays4)).strftime("%Y%m%d")
-    startAll = (curdate-timedelta(days=365)).strftime("%Y%m%d")
-    addDir(translation(30031) + str(replays1) +' Days',
-           dict(ESPN_URL=events.get_replay_events_url(channel_list) + enddate + '&startDate=' + start1, MODE=LIST_SPORTS_MODE),
-           defaultreplay)
-    addDir(translation(30031) + str(replays2) +' Days',
-           dict(ESPN_URL=events.get_replay_events_url(channel_list) + enddate + '&startDate=' + start2, MODE=LIST_SPORTS_MODE),
-           defaultreplay)
-    addDir(translation(30031) + str(replays3) +' Days',
-           dict(ESPN_URL=events.get_replay_events_url(channel_list) + enddate + '&startDate=' + start3, MODE=LIST_SPORTS_MODE),
-           defaultreplay)
-    addDir(translation(30031) + str(replays3) +'-' + str(replays4) +' Days',
-           dict(ESPN_URL=events.get_replay_events_url(channel_list) + '&endDate=' + start3 + '&startDate=' + start4, MODE=LIST_SPORTS_MODE),
-           defaultreplay)
-    addDir(translation(30032),
-           dict(ESPN_URL=events.get_replay_events_url(channel_list) + enddate + '&startDate=' + startAll, MODE=LIST_SPORTS_MODE),
-           defaultreplay)
-    xbmcplugin.endOfDirectory(pluginhandle)
-
-def LISTSPORTS(args):
-    espn_url = args.get(ESPN_URL)[0]
-    if 'action=replay' in espn_url:
-        image = defaultreplay
-    elif 'action=upcoming' in espn_url:
-        image = defaultupcoming
-    else:
-        image = defaultimage
-    addDir(translation(30034), dict(ESPN_URL=espn_url, MODE=LIVE_EVENTS_MODE), image)
-    sports = []
-    sport_elements = util.get_url_as_xml_soup_cache(espn_url).findall('.//sportDisplayValue')
-    for sport in sport_elements:
-        sport = sport.text.encode('utf-8')
-        if sport not in sports:
-            sports.append(sport)
-    for sport in sports:
-        addDir(sport, dict(ESPN_URL=espn_url, MODE=INDEX_SPORTS_MODE, SPORT=sport), image)
-    xbmcplugin.addSortMethod(pluginhandle, xbmcplugin.SORT_METHOD_VIDEO_SORT_TITLE)
-    xbmcplugin.endOfDirectory(pluginhandle)
-
-def INDEX_EVENT(event, live, upcoming, replay, chosen_sport):
-    sport = event.find('sportDisplayValue').text.encode('utf-8')
-    desktopStreamSource = event.find('desktopStreamSource').text
-    ename = event.find('name').text
-    eventid = event.get('id')
-    simulcastAiringId = event.find('simulcastAiringId').text
-    networkid = event.find('networkId').text
-    if networkid is not None:
-        network = player_config.get_network_name(networkid)
-    sport2 = event.find('sport').text
-    if sport <> sport2:
-        sport += ' ('+sport2+')'
-    league = event.find('league').text
-    location = event.find('site').text
-    fanart = event.find('.//thumbnail/large').text
-    fanart = fanart.split('&')[0]
-    mpaa = event.find('parentalRating').text
-    starttime = int(event.find('startTimeGmtMs').text)/1000
-    etime = time.strftime("%I:%M %p",time.localtime(float(starttime)))
-    endtime = int(event.find('endTimeGmtMs').text)/1000
-    start = time.strftime("%m/%d/%Y %I:%M %p",time.localtime(starttime))
-    aired = time.strftime("%Y-%m-%d",time.localtime(starttime))
-    udate = time.strftime("%m/%d",time.localtime(starttime))
-    now = datetime.now().strftime('%H%M')
-    etime24 = time.strftime("%H%M",time.localtime(starttime))
-    aspect_ratio = event.find('aspectRatio').text
-    length = str(int(round((endtime - time.time()))))
-    title_time = etime
-    if live and now > etime24:
-        color = 'FFFFFF'
-    elif live:
-        color = '999999'
-    else:
-        color = 'E0E0E0'
-        length = str(int(round((endtime - starttime))))
-        title_time = ' - '.join((udate, etime))
-
-    if network == 'longhorn':
-        channel_color = 'BF5700'
-    elif network == 'sec' or network == 'secplus':
-        channel_color = '004C8D'
-    else:
-        channel_color = 'CC0000'
-
-    ename = '[COLOR=FF%s]%s[/COLOR] [COLOR=FFB700EB]%s[/COLOR] [COLOR=FF%s]%s[/COLOR]' % (channel_color, network, title_time, color, ename)
-
-    length_minutes = int(length) / 60
-
-    end = event.find('summary').text
-    if end is None or len(end) == 0:
-        end = event.find('caption').text
-
-    if end is None:
-        end = ''
-    end += '\nNetwork: ' + network
-
-    plot = ''
-    if sport <> None and sport <> ' ':
-        plot += translation(30620) % sport + '\n'
-    if league <> None and league <> ' ':
-        plot += translation(30630) % league+'\n'
-    if location <> None and location <> ' ':
-        plot += translation(30640) % location+'\n'
-    if start <> None and start <> ' ':
-        plot += translation(30650) % start+'\n'
-    if length <> None and length <> ' ' and live:
-        plot += translation(30660) % str(length_minutes) + '\n'
-    elif length <> None and length <> ' ' and (replay or upcoming):
-        plot += translation(30670) % str(length_minutes) +'\n'
-    plot += end
-    infoLabels = {'title': ename,
-                  'genre':sport,
-                  'plot':plot,
-                  'aired':aired,
-                  'premiered':aired,
-                  'duration':length,
-                  'studio':network,
-                  'mpaa':mpaa,
-                  'videoaspect' : aspect_ratio}
-
-    session_url = 'http://broadband.espn.go.com/espn3/auth/watchespn/startSession?'
-    session_url += '&channel='+network
-    session_url += '&simulcastAiringId='+simulcastAiringId
-
-    authurl = dict()
-    authurl[EVENT_ID] = eventid
-    authurl[MODE] = UPCOMING_MODE if upcoming else PLAY_MODE
-    authurl[NETWORK_NAME] = event.find('adobeResource').text
-    authurl[EVENT_NAME] = event.find('name').text.encode('utf-8')
-    authurl[EVENT_GUID] = event.find('guid').text.encode('utf-8')
-    authurl[EVENT_PARENTAL_RATING] = event.find('parentalRating').text
-    authurl[SESSION_URL] = session_url
-    addLink(ename, authurl, fanart, fanart, infoLabels=infoLabels)
-
-def INDEX(args):
-    espn_url = args.get(ESPN_URL)[0]
-    chosen_sport = args.get(SPORT, None)
-    if chosen_sport is not None:
-        chosen_sport = chosen_sport[0]
-    chosen_network = args.get(NETWORK_ID, None)
-    if chosen_network is not None:
-        chosen_network = chosen_network[0]
-    live = 'action=live' in espn_url
-    upcoming = 'action=upcoming' in espn_url
-    replay = 'action=replay' in espn_url
-    if live:
-        data = events.get_events(espn_url)
-    else:
-        data = util.get_url_as_xml_soup_cache(espn_url).findall(".//event")
-    num_espn3 = 0
-    num_secplus = 0
-    num_events = 0
-    for event in data:
-        sport = event.find('sportDisplayValue').text.encode('utf-8')
-        if chosen_sport <> sport and chosen_sport is not None:
-            continue
-        networkid = event.find('networkId').text
-        if chosen_network <> networkid and chosen_network is not None:
-            continue
-        if networkid == ESPN3_ID and chosen_network is None and live :
-            num_espn3 = num_espn3 + 1
-        elif networkid == SECPLUS_ID and chosen_network is None and live :
-            num_secplus = num_secplus + 1
-        else:
-            num_events = num_events + 1
-            INDEX_EVENT(event, live, upcoming, replay, chosen_sport)
-    # Don't show ESPN3 folder if there are no premium events
-    if num_events == 0:
-        for event in data:
-            sport = event.find('sportDisplayValue').text.encode('utf-8')
-            if chosen_sport <> sport and chosen_sport is not None:
-                continue
-            INDEX_EVENT(event, live, upcoming, replay, chosen_sport)
-    # Dir for ESPN3/SECPlus
-    elif chosen_network is None:
-        if num_espn3 > 0:
-            translation_number = 30191 if num_espn3 == 1 else 30190
-            addDir('[COLOR=FFCC0000]' + (translation(translation_number) % num_espn3) + '[/COLOR]',
-               dict(ESPN_URL=espn_url, MODE=LIVE_EVENTS_MODE, NETWORK_ID=ESPN3_ID),
-               defaultlive)
-        if num_secplus > 0:
-            translation_number = 30201 if num_espn3 == 1 else 30200
-            addDir('[COLOR=FF004C8D]' + (translation(translation_number) % num_secplus) + '[/COLOR]',
-               dict(ESPN_URL=espn_url, MODE=LIVE_EVENTS_MODE, NETWORK_ID=SECPLUS_ID),
-               defaultlive)
-    xbmcplugin.setContent(pluginhandle, 'episodes')
-    xbmcplugin.endOfDirectory(pluginhandle)
-
 def check_blackout(item):
     blackouts = item.findall('.//blackouts/blackoutsItem/detail/detailItem')
     blackout_type = item.find('.//blackouts/blackoutsItem/type')
@@ -561,171 +315,6 @@ def PLAY_ITEM(args):
     url = args.get(PLAYBACK_URL)[0]
     item = xbmcgui.ListItem(path=url)
     return xbmcplugin.setResolvedUrl(pluginhandle, True, item)
-
-def check_error(session_json):
-    status = session_json['status']
-    if not status == 'success':
-        dialog = xbmcgui.Dialog()
-        dialog.ok(translation(30037), translation(30500) % session_json['message'])
-        return True
-    return False
-
-def does_requires_auth(network_name):
-    requires_auth = not network_name == 'espn3'
-    if network_name == 'espn3':
-        free_content_check = player_config.can_access_free_content()
-        if not free_content_check:
-            xbmc.log('ESPN3: User needs login to ESPN3', LOG_LEVEL)
-            requires_auth = True
-    return requires_auth
-
-
-def PLAY_LEGACY_TV(args):
-    # check blackout differently for legacy shows
-    event_id = args.get(EVENT_ID)[0]
-    url = 'http://broadband.espn.go.com/espn3/auth/watchespn/util/isUserBlackedOut?eventId=' + event_id
-    xbmc.log(TAG + 'Blackout url %s' % url, LOG_LEVEL)
-    blackout_data = util.get_url_as_json(url)
-    blackout = blackout_data['E3BlackOut']
-    if not blackout == 'true':
-        blackout = blackout_data['LinearBlackOut']
-    if blackout == 'true':
-        dialog = xbmcgui.Dialog()
-        dialog.ok(translation(30037), translation(30040))
-        return
-    PLAY_TV(args)
-
-# TODO: Unsure if cookie is needed
-#ua UA_PC
-#finalurl = finalurl + '|Connection=keep-alive&User-Agent=' + urllib.quote(ua) + '&Cookie=_mediaAuth=' + urllib.quote(base64.b64encode(pkan))
-def PLAY_TV(args):
-
-    network_name = args.get(NETWORK_NAME)[0]
-    event_name = args.get(EVENT_NAME)[0]
-    event_guid = args.get(EVENT_GUID)[0]
-    event_parental_rating = args.get(EVENT_PARENTAL_RATING)[0]
-    resource = adobe_activate_api.get_resource(network_name, event_name, event_guid, event_parental_rating)
-
-    requires_auth = does_requires_auth(network_name)
-
-    if requires_auth:
-        if not adobe_activate_api.is_authenticated():
-            dialog = xbmcgui.Dialog()
-            dialog.ok(translation(30037), translation(30410))
-            return
-        media_token = adobe_activate_api.get_short_media_token(resource)
-        token_type = 'ADOBEPASS'
-    else:
-        media_token = adobe_activate_api.get_device_id()
-        token_type = 'DEVICE'
-
-
-    # see http://api-app.espn.com/v1/watch/clients/watchespn-tvos for details
-    # see http://espn.go.com/watchespn/appletv/featured for details
-    start_session_url = args.get(SESSION_URL)[0]
-    params = urllib.urlencode({'partner':'watchespn',
-                               'playbackScenario':'HTTP_CLOUD_HIGH',
-                               'platform':'tvos',
-                               'token':media_token,
-                               'tokenType':token_type,
-                               'resource':base64.b64encode(resource),
-                               'v': '2.0.0'
-                               })
-    start_session_url += '&' + params
-
-    xbmc.log('ESPN3: start_session_url: ' + start_session_url, LOG_LEVEL)
-
-    session_json = util.get_url_as_json(start_session_url)
-    if check_error(session_json):
-        return
-
-    playback_url = session_json['session']['playbackUrls']['default']
-    stream_quality = str(selfAddon.getSetting('StreamQuality'))
-    bitrate_limit = int(selfAddon.getSetting('BitrateLimit'))
-    xbmc.log(TAG + 'Stream Quality %s' % stream_quality, LOG_LEVEL)
-    m3u8_obj = m3u8.load(playback_url)
-    success = True
-    if m3u8_obj.is_variant:
-        stream_options = list()
-        bandwidth_key = 'bandwidth'
-        m3u8_obj.playlists.sort(key=lambda playlist: playlist.stream_info.bandwidth, reverse=True)
-        m3u8_obj.data['playlists'].sort(key=lambda playlist: int(playlist['stream_info'][bandwidth_key]), reverse=True)
-        stream_quality_index = str(selfAddon.getSetting('StreamQualityIndex'))
-        stream_index = None
-        should_ask = False
-        try:
-            stream_index = int(stream_quality_index)
-            if stream_index < 0 or stream_index >= len(m3u8_obj.playlists):
-                should_ask = True
-        except:
-            should_ask = True
-        if '0' == stream_quality: # Best
-            stream_index = 0
-            should_ask = False
-            for playlist in m3u8_obj.data['playlists']:
-                stream_info = playlist['stream_info']
-                bandwidth = int(stream_info[bandwidth_key]) / 1024
-                if bandwidth <= bitrate_limit:
-                    break
-                stream_index = stream_index + 1
-        elif '2' == stream_quality: #Ask everytime
-            should_ask = True
-        if should_ask:
-            for playlist in m3u8_obj.data['playlists']:
-                stream_info = playlist['stream_info']
-                resolution = stream_info['resolution']
-                frame_rate = stream_info['frame_rate']
-                bandwidth = int(stream_info[bandwidth_key]) / 1024
-                if 'average_bandwidth' in stream_info:
-                    xbmc.log(TAG + 'bandwidth: %s average bandwidth: %s' %
-                             (stream_info['bandwidth'], stream_info['average_bandwidth']), LOG_LEVEL)
-                stream_options.append(translation(30450) % (resolution,
-                                                      frame_rate,
-                                                      bandwidth))
-            dialog = xbmcgui.Dialog()
-            stream_index = dialog.select(translation(30440), stream_options)
-            if stream_index < 0:
-                success = False
-            else:
-                selfAddon.setSetting(id='StreamQualityIndex', value=str(stream_index))
-
-        xbmc.log(TAG + 'Chose stream %d' % stream_index, LOG_LEVEL)
-        item = xbmcgui.ListItem(path=m3u8_obj.playlists[stream_index].uri)
-        return xbmcplugin.setResolvedUrl(pluginhandle, success, item)
-    else:
-        item = xbmcgui.ListItem(path=finalurl)
-        return xbmcplugin.setResolvedUrl(pluginhandle, success, item)
-
-def addLink(name, url, iconimage, fanart=None, infoLabels=None):
-    u = sys.argv[0] + '?' + urllib.urlencode(url)
-    liz = xbmcgui.ListItem(name, iconImage="DefaultVideo.png", thumbnailImage=iconimage)
-
-    if infoLabels is None:
-        infoLabels={"Title": name}
-    liz.setInfo(type="Video", infoLabels=infoLabels)
-    liz.setProperty('IsPlayable', 'true')
-    liz.setIconImage(iconimage)
-    if fanart is None:
-        fanart=defaultfanart
-    liz.setProperty('fanart_image',fanart)
-    video_streaminfo = dict()
-    liz.addStreamInfo('video', video_streaminfo)
-    ok = xbmcplugin.addDirectoryItem(handle=pluginhandle, url=u, listitem=liz)
-    return ok
-
-
-def addDir(name, url, iconimage, fanart=None, infoLabels=None):
-    u = sys.argv[0] + '?' + urllib.urlencode(url)
-    liz = xbmcgui.ListItem(name, iconImage="DefaultFolder.png", thumbnailImage=iconimage)
-    if infoLabels is None:
-        infoLabels={"Title": name}
-    liz.setInfo(type="Video", infoLabels=infoLabels)
-    if fanart is None:
-        fanart=defaultfanart
-    liz.setProperty('fanart_image',fanart)
-    ok = xbmcplugin.addDirectoryItem(handle=pluginhandle, url=u, listitem=liz, isFolder=True)
-    return ok
-
 
 base_url = sys.argv[0]
 args = urlparse.parse_qs(sys.argv[2][1:])
@@ -771,15 +360,15 @@ elif mode[0] == CATEGORY_SHOWCASE_MODE:
     CATEGORIES_SHOWCASE(args)
 elif mode[0] == LIVE_EVENTS_MODE:
     xbmc.log("Indexing Videos", LOG_LEVEL)
-    INDEX(args)
+    legacy.INDEX(args)
 elif mode[0] == LIST_SPORTS_MODE:
     xbmc.log("List sports", LOG_LEVEL)
-    LISTSPORTS(args)
+    legacy.LISTSPORTS(args)
 elif mode[0] == INDEX_SPORTS_MODE:
     xbmc.log("Index by sport", LOG_LEVEL)
-    INDEX(args)
+    legacy.INDEX(args)
 elif mode[0] == PLAY_MODE:
-    PLAY_LEGACY_TV(args)
+    legacy.PLAY_LEGACY_TV(args)
 elif mode[0] == PLAY_ITEM_MODE:
     PLAY_ITEM(args)
 elif mode[0] == PLAY_TV_MODE:
@@ -793,7 +382,7 @@ elif mode[0] == CATEGORY_SHELF_MODE:
     CATEGORY_SHELF(args)
 elif mode[0] == OLD_LISTING_MODE:
     xbmc.log("Old listing mode", LOG_LEVEL)
-    CATEGORIES()
+    legacy.CATEGORIES()
 elif mode[0] == CATEGORY_SPORTS_MODE:
     CATEGORY_SPORTS(args)
 elif mode[0] == CATEGORY_CHANNELS_MODE:
