+++ resources/lib/constants.py
@@ -1,0 +1,37 @@
+OLD_LISTING_MODE = 'OLD_LISTING_MODE'
+LIVE_EVENTS_MODE = 'LIVE_EVENTS'
+PLAY_MODE = 'PLAY'
+PLAY_ITEM_MODE = 'PLAY_ITEM'
+PLAY_TV_MODE = 'PLAY_TV'
+LIST_SPORTS_MODE = 'LIST_SPORTS'
+INDEX_SPORTS_MODE = 'INDEX_SPORTS'
+UPCOMING_MODE = 'UPCOMING'
+AUTHENTICATE_MODE = 'AUTHENTICATE'
+AUTHENTICATION_DETAILS_MODE = 'AUTHENTICATION_DETAILS'
+CATEGORY_SHELF_MODE = 'CATEGORY_SHELF'
+CATEGORY_SHOWCASE_MODE = 'CATEGORY_SHOWCASE'
+CATEGORY_SPORTS_MODE = 'CATEGORY_SPORTS'
+CATEGORY_CHANNELS_MODE = 'CATEGORY_CHANNELS'
+NETWORK_ID = 'NETWORK_ID'
+EVENT_ID = 'EVENT_ID'
+SIMULCAST_AIRING_ID = 'SIMULCAST_AIRING_ID'
+SESSION_URL = 'SESSION_URL'
+DESKTOP_STREAM_SOURCE = 'DESKTOP_STREAM_SOURCE'
+NETWORK_NAME = 'NETWORK_NAME'
+EVENT_NAME = 'EVENT_NAME'
+EVENT_GUID = 'EVENT_GUID'
+EVENT_PARENTAL_RATING = 'EVENT_PARENTAL_RATING'
+SHELF_ID = 'SHELF_ID'
+SHOWCASE_URL = 'SHOWCASE_URL'
+SHOWCASE_NAV_ID = 'SHOWCASE_NAV_ID'
+PLAYBACK_URL = 'PLAYBACK_URL'
+
+ESPN_URL = 'ESPN_URL'
+MODE = 'MODE'
+SPORT = 'SPORT'
+
+BAM_NS = '{http://services.bamnetworks.com/media/types/2.1}'
+
+# Taken from https://espn.go.com/watchespn/player/config
+ESPN3_ID = 'n360'
+SECPLUS_ID = 'n323'
